+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Intro, start and game over scroll issues are fixed.
 - Default game duration is 5 minutes now.
 - List of incorrect operations is shown at the end of the game.
+- Typing digits is disallowed when challenge has an answer.
 
 ## 0.1.0 (2020-11-30)